Продолжает 12 задание в этой ветке
