+++ js/main.js
@@ -2,6 +2,7 @@ import './make-pictures.js';
 import './make-bigpicture.js';
 import './upload-photo.js';
 import './scale.js';
+import './validation-form.js';
 // eslint-disable-next-line no-unused-vars
 import { createMockPhotos } from './data.js';
 //console.log(createMockPhotos());
