+++ js/upload-photo.js
@@ -1,6 +1,4 @@
 import { isEscapeKey } from './util.js';
-import { sendData } from './api.js';
-import {createSuccess, createError} from './success.js';
 
 
 const bodyElement = document.querySelector('body');
@@ -9,22 +7,6 @@ const bodyElement = document.querySelector('body');
 const imgInputElement = document.querySelector('#upload-file');
 const overlayElement = document.querySelector('.img-upload__overlay');
 const cancelButton = document.querySelector('#upload-cancel');
-const uploadForm = document.querySelector('.img-upload__form');
-const submitButtonElement = uploadForm.querySelector('#upload-submit');
-
-
-const HASHTAG_REGEX = /#[a-zа-яё0-9]{1,19}$/i;
-const MAX_HASHTAG_COUNT = 5;
-
-const texthashtagElement = uploadForm.querySelector('.text__hashtags');
-const textDescriptionElement = uploadForm.querySelector('.text__description');
-
-const pristine = new Pristine(uploadForm, {
-  classTo: 'img-upload__field-wrapper',
-  errorClass: 'form__item--invalid',
-  errorTextParent: 'img-upload__field-wrapper',
-  errorTextClass: 'text__hashtags-error',
-});
 
 imgInputElement.addEventListener('change',() => {
   showModal();
@@ -41,10 +23,12 @@ const onDocumentKeydown = (evt) => {
 
 };
 
+
 function showModal() {
   overlayElement.classList.remove('hidden');
   bodyElement.classList.add('modal-open');
   document.addEventListener('keydown', onDocumentKeydown);
+
 
 }
 
@@ -52,114 +36,9 @@ function hideModal() {
   overlayElement.classList.add('hidden');
   bodyElement.classList.remove('modal-open');
   document.removeEventListener('keydown', onDocumentKeydown);
-  imgInputElement.value = '';
+
 }
 cancelButton.addEventListener('click', () =>
   hideModal()
 );
-
-
-const validateHashtag = (text) => HASHTAG_REGEX.test(text) || text === '';
-
-
-const validateHashtagCount = (tags) =>
-  tags
-    .split('')
-    .filter((tag) => tag === '#')
-    .length <= MAX_HASHTAG_COUNT;
-
-
-const validateUniqHashtags = (tag) => {
-  const tagArray = tag
-    .replaceAll(' ','')
-    .toLowerCase()
-    .split('#');
-  tagArray.shift();
-
-  const unique = Array.from(new Set(tagArray));
-
-  return tagArray.length === unique.length;
-
-};
-
-
-// Валидатор правильности хештега
-pristine.addValidator(
-  texthashtagElement,
-  validateHashtag,
-  'Ошибка! не верно введен хештег'
-);
-
-// Валидатор на количество хештегов
-pristine.addValidator(
-  texthashtagElement,
-  validateHashtagCount,
-  'Ошибка! максимальное количество хештегов: 5'
-);
-
-// Валидатор на одинаковые хештеги
-pristine.addValidator(
-  texthashtagElement,
-  validateUniqHashtags ,
-  'Ошибка! Одинаковые хештеги!'
-);
-
-
-const blockSubmitButton = () => {
-  submitButtonElement.disabled = true;
-  submitButtonElement.textContent = 'Загружаю...';
-
-};
-
-const unblockSubmitButton = () => {
-  submitButtonElement.disabled = false;
-  submitButtonElement.textContent = 'Опубликовать';
-
-};
-
-const setUserFormSubmit = (onSuccess) => {
-  uploadForm.addEventListener('submit', (evt) => {
-    evt.preventDefault();
-    pristine.validate();
-    const isValid = pristine.validate();
-    if (isValid) {
-      blockSubmitButton();
-      sendData(
-        () => {
-          onSuccess();
-          blockSubmitButton();
-          createSuccess();
-          bodyElement.classList.add('modal-open');
-        },
-        () => {
-          createError();
-          unblockSubmitButton();
-          bodyElement.classList.add('modal-open');
-        },
-        new FormData(evt.target),
-        unblockSubmitButton
-      );
-    }
-
-  });
-};
-
-texthashtagElement.addEventListener('keydown', (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.stopPropagation();
-    document.activeElement.blur();
-  }
-});
-
-textDescriptionElement.addEventListener('keydown', (evt) => {
-  if (isEscapeKey(evt)) {
-    evt.stopPropagation();
-    document.activeElement.blur();
-  }
-});
-
-
-setUserFormSubmit(hideModal);
-
-
 export {onDocumentKeydown, hideModal};
